Cast has_wait_time expression so pgx can infer the parameter type
$5 used both as a column value and in "is not null" trips
"could not determine data type of parameter" (42P08) under pgx's
prepared batches; an explicit ::int cast resolves it. Verified
live on the VPS: poll ok after redeploy.

+++ service/internal/store/store.go
@@ -66,7 +66,7 @@ const readingInsertSQL = `
 insert into wait_time_readings (
   location_id, report_id, reading_id, reading_created_at,
   wait_time_minutes, elos_minutes, status, has_wait_time
-) values ($1,$2,$3,$4,$5,$6,$7,$5 is not null)
+) values ($1,$2,$3,$4,$5,$6,$7,$5::int is not null)
 on conflict (location_id, report_id) do nothing
 returning id`
